% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_geo.R
\name{ag_call}
\alias{ag_call}
\title{Internal utility}
\usage{
ag_call(data = w_com_cp, by = by_w)
}
\arguments{
\item{data}{: données}

\item{by}{: by}
}
\value{
tibble
}
\description{
Internal utility
}
\author{
Edouard Chatignoux
}
\keyword{internal}
