% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_iris.R
\name{com_to_iris}
\alias{com_to_iris}
\title{Pass from com to iris geography}
\usage{
com_to_iris(out)
}
\arguments{
\item{out}{: com to com correspondance table}
}
\value{
A tibble of correspondance for the iris levels
}
\description{
Internal function that links IRIS to communes from one geography
to another geography.
Correspondance between IRIS and communes is one-one.
}
\author{
Edouard Chatignoux
}
\keyword{internal}
