% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_geo.R
\name{cp_to_geo}
\alias{cp_to_geo}
\title{Internal function to ventilate postal code to
commune/IRIS code}
\usage{
cp_to_geo(data, by = ~1, geo = "iris")
}
\arguments{
\item{data}{: table with postal codes}

\item{by}{: formula to compute weights, i.e. use com/IRIS age and/or sex structure to ventilate cases}

\item{geo}{: out geography ("iris" or "com")}
}
\value{
A tibble
}
\description{
Note that population estimates from 2017 (geography 2019) is used to ventilate cases
}
\author{
Edouard Chatignoux
}
