% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{get_data}
\alias{get_data}
\title{Extract data from the SQLite database dts_com}
\usage{
get_data(dt)
}
\arguments{
\item{dt}{data in the SQLite database \code{\link{dts_com.sqlite}}}
}
\value{
A tibble
}
\description{
Extract data from the SQLite database dts_com
}
\section{Eight data sets available}{

\describe{
\itemize{
\item \code{pop_com}: communes' population size by one year age and sex for years 2006 to 2018
\item \code{pass_com}: full correpondance of communes Id for years 2003 to 2021
\item \code{pass_iris}: full correpondance of iris Id for years 2016 to 2021
\item \code{fus_com}: data listing commune's fusions
\item \code{scis_com}: data listing commune's scissions
\item \code{p_cp_iris}: probability for an observation from postal code to belong to an iris (P(iris|cp))
\item \code{p_cp_com}: probability for an observation from postal code to belong to an commune (P(com|cp))
\item \code{ref_geo}: reference table for iris and comunes for years 2008 to 2020
}
}
}

\examples{
get_data(pass_com)
}
\seealso{
\code{\link{dts_com.sqlite}}
}
\author{
Edouard Chatignoux
}
