% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_com.R
\name{pass_com.w.one}
\alias{pass_com.w.one}
\title{Fonction interne pour calculer les poids de passages d'une année sur l'autre}
\usage{
pass_com.w.one(
  geo = 2016,
  forward = TRUE,
  by = ~cut(age, breaks = c(-Inf, 50, Inf)) + sexe,
  dt_pop = NULL
)
}
\arguments{
\item{geo}{: année de départ}

\item{forward}{: TRUE si on augmente de millésime, FALSE sinon}

\item{by}{: strates de pop utilisées pour le calcul des poids (peut contenir des formules de type cut)}

\item{dt_pop}{: population (or weigth) table to use (default table if NULL)}
}
\value{
Une table de poids
}
\description{
Fonction interne pour calculer les poids de passages d'une année sur l'autre
}
\examples{
pass_com.w.one(geo = 2016,by=~sexe)
}
\seealso{
\code{\link{pass_com}}, \code{\link{dts_com.sqlite}}
}
\author{
Edouard Chatignoux
}
\keyword{internal}
