% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_geo.R
\name{pass_geo}
\alias{pass_geo}
\title{Fonction de passage d'une géographie à une autre}
\usage{
pass_geo(
  geo_in = 2016,
  geo_out = 2020,
  by = NULL,
  data = NULL,
  data_by = TRUE,
  geo = ~com
)
}
\arguments{
\item{geo_in}{Millésime des codes communes (i.e. ceux de data si
data non NULL)}

\item{geo_out}{Millésime des codes communes souhaités en sortie}

\item{by}{: strates de pop utilisées pour le calcul des poids
(peut contenir des formules, de type cut par example); si des
variables de data doivent être conservées telle que, les mettre
dans le by (voir exemples)}

\item{data}{Table de données dans la géographie millésimée geo_in
à sortir en géographie millésimée geo_out. La table doit avoir
des colonnes compatibles avec le by.  Si laissée à NULL, la
table des poids de passage est retournée.}

\item{data_by}{: Mettre a TRUE pour regrouper data selon by}

\item{geo}{: Formule donnant le niveau géographique d'entrée et de
sortie souhaités : \itemize{ \item \code{cp~iris}: pour cp en
entrée et iris en sortie \item \code{com~iris}: pour com en
entrée et iris en sortie \item \code{com~com}: pour com en
entrée et en sortie (idem à \code{~com}) } Toutes les
combinaisons sont possibles, sauf avec le cp en sortie et l'iris
en entrée (\code{com~cp} et \code{iris~com} sont inopérants par
exemple).}
}
\value{
Une table dans le nouveau millésime des communes
}
\description{
Par défaut, la fonction prend en argument une table d'entrée avec une définition des
communes dans un millésime donné et sort une table dans un autre millésime.
Les données (variables numériques) des communes fusionnées entre les deux millésime sont aggrégées.
Les données (variables numériques) des communes séparées entre les deux millésime sont réparties entre les
nouvelles communes au prorata de leur population lors de la première années de la scission.
Les strates de population utilisées pour calculer ce prorata est laissé au choix de l'utilisateur.
Si aucune table d'entrée n'est spécifiée, une table de conversion entre les géographie est retournée.
Si l'option géo est modifiée, la fonction permet le passage d'un code postal vers des communes/iris, et des communes vers les iris (mais ce n'est certainement pas souhaitable).
}
\examples{
## Pour avoir les poids de passage
## uniquement
pass_geo(geo_in=2015,
         geo_out=2017,
         by=~1)

## Pour appliquer le passage  directement à une table
## Ex. table des pops géo 2013 à passer en version géo 2017
## Poids calculés par classe d'age * sexe
## On veut garder les infos contenues dans les colonnes dep et annee_pop
data(dt_pop_2013)
pass_geo(data=dt_pop_2013,
         geo_in=2013,
         geo_out=2017,
         by=~cut(age,breaks=c(-Inf,50,Inf))+sexe+annee_pop+dep,
         geo=  ~ com)
}
\author{
Edouard Chatignoux
}
